refactor: 重构域名证书部署和证书管理，将原来代理维度部署改为域名为维度，解决多域名代理共享同一证书不匹配问题；重构SSL相关UI

+++ console/src/main/java/com/xiaoniucode/etp/server/web/repository/ProxyDomainRepository.java
@@ -12,7 +12,7 @@ import java.util.Optional;
  * 代理域名 Repository
  */
 @Repository
-public interface ProxyDomainRepository extends JpaRepository<ProxyDomainDO, Integer> {
+public interface ProxyDomainRepository extends JpaRepository<ProxyDomainDO, Long> {
 
     void deleteByProxyId(String proxyId);
